forex-choch-detector/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forex-choch-detector"
version = "1.0.0"
description = "Production-grade real-time CHoCH detection system"
requires-python = ">=3.11"
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "pydantic>=1.10.0",
    "PyYAML>=6.0",
    "structlog>=22.0.0",
    "aiohttp>=3.8.0",
    "typer>=0.7.0",
    "rich>=12.0.0",
    "fastapi>=0.85.0"
]

[project.scripts]
forex-choch = "src.cli.main:app"

[tool.black]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
